+++ TreeAna/g_clx.cc
@@ -46,7 +46,7 @@ void g_clx::Loop( string outputfilename ) {
 
 	// Ratio of prompt and random time windows
 	// Alternatively, normalisation of beta-decay lines
-	double bg_frac = -1.0;
+	float bg_frac = -1.0;
 
 	// Test if it's an electron or gamma
 	bool electron;
